+++ src/Controllers/HeaderController.php
@@ -15,5 +15,40 @@
 				$this->response->data["javascript"] ?? [],
 				$this->template->assets->getJavascript("header")
 			);
+
+			$this->response->data["bodyClasses"] = $this->getBodyTagClasses();
+		}
+
+		/**
+		 * Determines classes suiteable for the <body> tag
+		 * These classes can be used for easier identification of controller and view files used
+		 * or CSS styling for specific conditions
+		 * 
+		 * @return string An escaped string suitable for printing to the 'class' attribute of the <body> tag
+		 */
+		private function getBodyTagClasses() : string {
+			$controllerName = $this->getApplication()->getExecutedClassName()->toStringWithoutSuffix();
+			$methodName 	= $this->getApplication()->getCalledMethodName()->toStringWithoutSuffix();
+
+			$bodyClasses = [];
+			$bodyClasses[] = $controllerName;
+			$bodyClasses[] = $controllerName . '-' . $methodName;
+			$bodyClasses[] = $this->getResponse()->getView();
+
+			foreach($this->getChildren() as $childControllerName) {
+				$bodyClasses[] = $childControllerName->toStringWithoutSuffix();
+			}
+
+			foreach($this->getRequest()->getArguments() as $arg) {
+				$bodyClasses[] = $arg;
+			}
+
+			foreach($bodyClasses as $i => $bodyClass) {
+				$bodyClasses[$i] = strtolower($bodyClass);
+			}
+
+			$classesString = implode(' ', array_unique($bodyClasses));
+
+			return \HtmlEscape::escape($classesString);
 		}
 	}